requirements.txt
# LLM y embeddings
openai>=1.0.0

# Base de datos vectorial
qdrant-client>=1.7.0

# Parser PDF
pymupdf>=1.23.0

# Interfaz de usuario
gradio>=4.0.0

# Utilidades
python-dotenv>=1.0.0

# Procesamiento de imágenes y tokens
pillow>=10.0.0
tiktoken>=0.5.0
